% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumen_numerico.R
\name{resumen_numerico}
\alias{resumen_numerico}
\title{Resumen Numerico}
\usage{
resumen_numerico(datos, columna)
}
\arguments{
\item{datos}{Archivo de datos}

\item{columna}{Columna de de datos de tipo numerico de la cual se quiere obtenr un resumen numerico}
}
\value{
Devuelve un tibble o tabla con el maximo, minimo, promedio,
desvio estandar y mediana de una columna del archvo que se le pasa como argumento
}
\description{
La funcion resumen_numerico sirve para obtener una tabla o tibble con el maximo, minimo, promedio,
desvio estandar y mediana de una columna del archvo que se le pasa como argumento
}
\examples{
resumen_numerico(pinguinos, alto_pico_mm)

}
